==> project_code/01_MR_related/scripts/legacy_V2/02_explore_mr_results.Rmd <==
---
title: "Explore traits with effect in at least one outcome"
output: html_notebook
---

```{r message=F}
library(dplyr)
library(readr)
library(tidyr)
library(stringr)
library(ggplot2)
library(plotly)
library(wesanderson)
library(RColorBrewer)
library(flextable)
library(TwoSampleMR)
source('app1_MR-EvE_app/functions.R')
```


```{r, message = FALSE, include=F}
rm(dat)
dat <- #read_tsv("explore_MR-EvE_app/data_copy/bc_all_mr.tsv") %>% 
        #read_tsv("explore_MR-EvE_app/data_copy/bc_all_mr_madewR.tsv") %>% 
      #read_tsv("app1_MR-EvE_app/data_copy/bc_all_mr_fromCIs.tsv") %>%  # from 01_mr_epigraph_query.R script
      read_tsv("app1_MR-EvE_app/data_copy/bc_all_mr_fromCIsV3.tsv") %>% 
      # subset 
      filter(exposure.sex != 'Males') %>% 
      #filter(mr.method != 'Steiger null') %>%  # NA results --- does not make difference to the total number of traits
      # convert MR results to OR
      tidy_display_numbers()%>% 
      # deal with al outcome related changes
      process_bc_outcomes() %>% 
      # create categories of exposure traits
      filter(!grepl("_raw",exposure.id)) %>% 
      create_exposure_categories() %>% 
      add_exposure_labels() 

length(unique(dat$exposure.id)) # 1643 # 321
dat %>% select(exposure.id, exposure.trait, exposure_cat) %>% distinct() %>% View()

```

```{r}
outcome_table<- dat %>% 
  mutate(case_percent = round(N_case/outcome.sample_size*100, 2)) %>% 
  select(outcome, outcome.id, outcome.sample_size, N_case, case_percent, chip, outcome.year, outcome.nsnp)%>% 
  mutate(chip=gsub("_","", chip)) %>% 
  distinct() %>%
  arrange(outcome, desc(outcome.sample_size)) %>% 
  rename(`sample_size`=outcome.sample_size)


set_flextable_defaults(big.mark = " ", 
  font.size = 10, theme_fun = theme_vanilla,
  padding.bottom = 6, 
  padding.top = 6,
  padding.left = 6,
  padding.right = 6,
  background.color = "#EFEFEF")

ft<-flextable(outcome_table)
ft<-width(ft, j = 1, width=1.5)
ft<-width(ft, j = 2, width=1.7)
ft


```

```{r}

exposure_table<- dat %>% 
  select(exposure_cat, exposure.trait, exposure.id) %>% 
  distinct() %>% 
  count(exposure_cat) %>% arrange(-n)


set_flextable_defaults(big.mark = " ", 
  font.size = 10, theme_fun = theme_vanilla,
  padding.bottom = 6, 
  padding.top = 6,
  padding.left = 6,
  padding.right = 6,
  background.color = "#EFEFEF")

ft<-flextable(exposure_table)
ft

keep <- 
c('Proteins',                      
'Metabolites',          
'Antrophometric',        
'Other biomarkers',      
'Diet and supplements',  
'Drugs',                 
'Physical activity',     
'Smoking',                    
'Reproductive',                      
'Alcohol',               
'Sleep')


dat %>% 
  filter(exposure_cat %in% keep) %>% 
  filter(chip %in% c('Meta', 'iCOG2017', 'OncArray')) %>% 
  select(exposure.id, outcome.id, mr.method) %>% 
  distinct() %>% 
  count(mr.method) %>% arrange(-n) %>% 
  mutate(pct =  round(n/sum(n)*100, 2)) 


```


```{r}
# output df will be saved here
tidy_traits_by_cat <- data.frame()
```


### 1. Anthophometric traits

```{r fig.width=10}
antro_blacklist <- c('ieu-a-81','ieu-a-74', "ieu-a-73" ,"ieu-a-79" ,"ieu-a-72" ,"ieu-a-78",
                     'ieu-a-63',  'ieu-a-66', 'ieu-a-60',  'ieu-a-69' , 'ieu-a-61',
                      'ieu-a-54', 'ieu-a-55',  'ieu-a-49' , 'ieu-a-48', 'ieu-a-57' , 'ieu-a-50',
                     'ukb-b-12039', 'ukb-b-2303',
                     'ieu-a-2', 'ieu-a-835')


input <- dat %>% filter(exposure_cat %in% c('Antrophometric')) %>% 
    filter(!grepl("arm|leg|first child", exposure.trait, ignore.case = T)) %>% 
    filter(!exposure.id %in% antro_blacklist)

length(unique(input$exposure.id)) #56

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


# deal with duplicated from UKB
ukb_diff_sources<-input %>%
        select(exposure.trait, exposure.id, author, consortium, exposure.sample_size, year) %>% 
        filter(author %in% c("Neale", "Ben Elsworth")) %>% distinct() %>% 
        count(exposure.trait) %>% filter(n==2) 

input <- input %>% 
  filter(!(exposure.trait %in% ukb_diff_sources$exposure.trait & author == 'Neale'))
length(unique(input$exposure.id)) # 53


antro_exclusion_table <- dat %>% 
                filter(exposure_cat %in% c('Antrophometric')) %>% 
                mutate(include = case_when(grepl("arm|leg|first child", exposure.trait, ignore.case = T) ~ F,
                                           exposure.id %in% antro_blacklist ~ F,
                                           exposure.trait %in% ukb_diff_sources$exposure.trait & author == 'Neale' ~ F,
                                           TRUE ~ T)) %>% 
                select(exposure.id, exposure.trait,exposure, include ) %>% distinct()

write_csv(antro_exclusion_table, "../results/mr_evidence_outputs/antro_traits_exclusion_list.csv")




p<-plot_bubble_plot(input)
ply<-plotly::ggplotly(p)


p_gg<-plotly::plotly_build(p) # convert ggplot object to plotly
p_gg$layout$showlegend = "TRUE"

```




### 2. Activity

```{r fig.width=10, fig.height=5}
input <- dat %>% filter(exposure_cat %in% c('Physical activity')) %>% 
             filter(!grepl("leisure|mental", exposure, ignore.case = T))
length(unique(input$exposure.id)) #39
#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)

```

### 3.  Supplements / diet

```{r fig.width=10, fig.height=5}
input <- dat %>% filter(exposure_cat %in% c('Diet and supplements')) %>% 
                filter(!grepl("questionnaire", exposure))

length(unique(input$exposure.id)) #54
#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)

```


### 4. Reproductive traits

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Reproductive')) 
length(unique(input$exposure.id)) # 23

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)

```



### 5. alcohol

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Alcohol')) 
length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)

```

### 6. smoking traits

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c( "Smoking")) 
length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)

```



### 7. Metabolites


```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Metabolites')) 

length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)



p<-plot_bubble_plot(input)


test_list<-input %>%  filter(mr.pval < 10e-6) %>% 
                      #filter(mr.b > 0.01) %>%
                      pull(exposure.id) %>% unique()
length(test_list)

p<-plot_bubble_plot(input %>% filter(exposure.id %in% test_list))
ply<-plotly::ggplotly(p)

```




### 8. Protein measures

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Proteins')) %>% 
        arrange(exposure.trait)

length(unique(input$exposure.id)) 
#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)



# testing visualization
dim(input)[1]/9
length(unique(input$exposure.trait))

p<-plot_bubble_plot(input[1:1000,])
p<-plot_bubble_plot(input[1000:2000,])

test_list<-input %>%  filter(mr.pval < 10e-8) %>% filter(mr.b > 0.01) %>% pull(exposure.id) %>% unique()
length(test_list)

p<-plot_bubble_plot(input %>% filter(exposure.id %in% test_list[1:60]))
p<-plot_bubble_plot(input %>% filter(exposure.id %in% test_list[61:131]))


p<-plot_bubble_plot(input %>% filter(grepl('interl', exposure.trait, ignore.case = T)))

ply<-plotly::ggplotly(p)

```

### 9. Other biomarkers/compounds

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Other biomarkers'))
length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)
```



### 10. Sleep

```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Sleep'))
length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


p<-plot_bubble_plot(input)
```

### 11. Drugs
```{r fig.width=12}
input <- dat %>% filter(exposure_cat %in% c('Drugs')) 
length(unique(input$exposure.id)) 
#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)

p<-plot_bubble_plot(input)
```



### 12. Lipids


```{r fig.width=10, fig.height=3}
input <- dat %>% filter(exposure_cat %in%  c('Lipids')) 

length(unique(input$exposure.id)) 

#save
tidy_traits_by_cat<-bind_rows(tidy_traits_by_cat, input)


```


```{r}
#SAVE
tidy_traits_by_cat %>% select(exposure.id, exposure_cat) %>% distinct() %>%  count(exposure_cat, sort=T)
tidy_traits_by_cat %>% select(exposure.id, exposure_cat) %>% distinct() %>% count(exposure_cat) %>% summarise(across(n, sum)) # 905
tidy_traits_by_cat %>% write_tsv("../results/mr_evidence_outputs/tidy_traits_by_cat.tsv")
```



## checking 
```{r}
old <- read_tsv("../results/mr_evidence_outputs/tidy_traits_by_cat_old.tsv")
old %>% select(exposure.id, exposure_cat) %>% distinct() %>%  count(exposure_cat, sort=T)
old %>% select(exposure.id, exposure_cat) %>% distinct() %>% count(exposure_cat) %>% summarise(across(n, sum))
```

```{r}
missing<- setdiff(old$exposure.id, tidy_traits_by_cat$exposure.id)
old %>% filter(exposure.id %in% missing) %>% select(exposure.id, exposure.trait, exposure, exposure_cat) %>% distinct() %>% View()
# basically just weird metabolites and cell types
```
